+++ e2e/colorpicker.e2e-spec.ts
@@ -21,8 +21,8 @@ describe('Colorpicker', () => {
   });
 
   it('should match previous opened screenshot', (done) => {
-    element(by.css('sky-dropdown button')).click();
-    expect('#screenshot-colorpicker2').toMatchBaselineScreenshot(done, {
+    element(by.css('.sky-colorpicker-button')).click();
+    expect('#screenshot-colorpicker-opened').toMatchBaselineScreenshot(done, {
       screenshotName: 'colorpicker-opened'
     });
   });
@@ -41,8 +41,8 @@ describe('Colorpicker (small screens)', () => {
   });
 
   it('should match previous opened screenshot', (done) => {
-    element(by.css('sky-dropdown button')).click();
-    expect('#screenshot-colorpicker2').toMatchBaselineScreenshot(done, {
+    element(by.css('.sky-colorpicker-button')).click();
+    expect('#screenshot-colorpicker-opened').toMatchBaselineScreenshot(done, {
       screenshotName: 'colorpicker-opened-xs'
     });
   });