+++ src/app/public/modules/colorpicker/colorpicker-input.directive.ts
@@ -3,13 +3,12 @@ import {
   ElementRef,
   forwardRef,
   HostListener,
+  Injector,
   Input,
   OnChanges,
   OnDestroy,
   OnInit,
-  Renderer,
-  SimpleChanges,
-  Injector
+  Renderer2
 } from '@angular/core';
 
 import {
@@ -39,9 +38,12 @@ import {
 } from './colorpicker.component';
 
 import {
-  SkyColorpickerHsva,
+  SkyColorpickerHsva
+} from './types/colorpicker-hsva';
+
+import {
   SkyColorpickerOutput
-} from './types';
+} from './types/colorpicker-output';
 
 // tslint:disable:no-forward-ref no-use-before-declare
 const SKY_COLORPICKER_VALUE_ACCESSOR = {
@@ -108,7 +110,7 @@ export class SkyColorpickerInputDirective
 
   constructor(
     private elementRef: ElementRef,
-    private renderer: Renderer,
+    private renderer: Renderer2,
     private service: SkyColorpickerService,
     private resourcesService: SkyLibResourcesService,
     private injector: Injector
@@ -131,7 +133,7 @@ export class SkyColorpickerInputDirective
 
   /* istanbul ignore next */
   @HostListener('blur')
-  public onBlur(event: any) {
+  public onBlur() {
     /*istanbul ignore next */
     this._onTouched();
   }
@@ -139,7 +141,7 @@ export class SkyColorpickerInputDirective
   public ngOnInit() {
     const element = this.elementRef.nativeElement;
 
-    this.renderer.setElementClass(element, 'sky-form-control', true);
+    this.renderer.addClass(element, 'sky-form-control');
     this.skyColorpickerInput.initialColor = this.initialColor;
     this.skyColorpickerInput.returnFormat = this.returnFormat;
 
@@ -155,9 +157,9 @@ export class SkyColorpickerInputDirective
     this.skyColorpickerInput.setColorFromString(this.initialColor);
 
     /// Set aria-label as default, if not set
-    if (!this.elementRef.nativeElement.getAttribute('aria-label')) {
-      this.renderer.setElementAttribute(
-        this.elementRef.nativeElement,
+    if (!element.getAttribute('aria-label')) {
+      this.renderer.setAttribute(
+        element,
         'aria-label',
         this.getString('skyux_colorpicker_input_default_label'));
     }
@@ -170,7 +172,7 @@ export class SkyColorpickerInputDirective
     }
 
     element.setAttribute('readonly', 'true');
-    this.renderer.setElementClass(element, 'sky-colorpicker-input', true);
+    this.renderer.addClass(element, 'sky-colorpicker-input');
   }
 
   public ngOnDestroy() {
@@ -189,7 +191,7 @@ export class SkyColorpickerInputDirective
     );
   }
 
-  public ngOnChanges(changes: SimpleChanges): void {
+  public ngOnChanges(): void {
     this._validatorChange();
     this.skyColorpickerInput.returnFormat = this.returnFormat;
     this.setColorPickerDefaults();
@@ -209,7 +211,7 @@ export class SkyColorpickerInputDirective
         this.skyColorpickerInput.initialColor = value;
       }
       this.skyColorpickerInput.lastAppliedColor = value;
-      let control: FormControl = (<NgControl>this.injector.get(NgControl)).control as FormControl;
+      let control: FormControl = (this.injector.get<NgControl>(NgControl)).control as FormControl;
       if (control) {
         control.setValue(this.modelValue, { emitEvent: false });
       }
@@ -250,8 +252,8 @@ export class SkyColorpickerInputDirective
 
     this.skyColorpickerInput.setColorFromString(output);
 
-    this.renderer.setElementStyle(element, 'background-color', setElementValue);
-    this.renderer.setElementProperty(element, 'value', output);
+    this.renderer.setStyle(element, 'background-color', setElementValue);
+    this.renderer.setProperty(element, 'value', output);
   }
 
   private formatter(color: any) {