+++ src/app/public/modules/colorpicker/colorpicker.module.ts
@@ -7,15 +7,18 @@ import {
 } from '@angular/common';
 
 import {
-  SkyDropdownModule
-} from '@skyux/popovers';
+  SkyAffixModule,
+  SkyCoreAdapterModule,
+  SkyOverlayModule
+} from '@skyux/core';
+
 import {
   SkyI18nModule
 } from '@skyux/i18n';
 
 import {
   SkyColorpickerResourcesModule
-} from '../shared';
+} from '../shared/colorpicker-resources.module';
 
 import {
   SkyColorpickerComponent
@@ -42,9 +45,11 @@ import {
   ],
   imports: [
     CommonModule,
+    SkyAffixModule,
     SkyColorpickerResourcesModule,
+    SkyCoreAdapterModule,
     SkyI18nModule,
-    SkyDropdownModule
+    SkyOverlayModule
   ],
   exports: [
     SkyColorpickerComponent,