+++ CHANGELOG.md
@@ -1,3 +1,51 @@
+# 4.0.2 (2020-06-22)
+
+- Removed unused parameters from the colorpicker input directive `HostListener` to address build issues when consumed by another library. [#48](https://github.com/blackbaud/skyux-colorpicker/pull/48)
+
+# 4.0.1 (2020-06-05)
+
+- Fixed the exports API to include `SkyColorpickerChangeAxis`, `SkyColorpickerCmyk`, `SkyColorpickerChangeColor`, `SkyColorpickerHsla`, `SkyColorpickerHsva`, and `SkyColorpickerRgba`. [#47](github.com/blackbaud/skyux-colorpicker/pull/47)
+
+# 4.0.0 (2020-05-21)
+
+### New features
+
+- Updated the pipeline to transpile to the [Angular Package Format](https://docs.google.com/document/d/1CZC2rcpxffTDfRDs6p1cfbmKNLA6x5O-NtkJglDaBVs/preview). [#43](https://github.com/blackbaud/skyux-colorpicker/pull/43)
+
+### Breaking changes
+
+- Dropped support for `rxjs@5`. Consumers can install `rxjs-compat@^6` to support older versions of `rxjs`. [#43](https://github.com/blackbaud/skyux-colorpicker/pull/43)
+
+# 4.0.0-rc.0 (2020-04-17)
+
+### New features
+
+- Updated the pipeline to transpile to the [Angular Package Format](https://docs.google.com/document/d/1CZC2rcpxffTDfRDs6p1cfbmKNLA6x5O-NtkJglDaBVs/preview). [#43](https://github.com/blackbaud/skyux-colorpicker/pull/43)
+
+### Breaking changes
+
+- Dropped support for `rxjs@5`. Consumers can install `rxjs-compat@^6` to support older versions of `rxjs`. [#43](https://github.com/blackbaud/skyux-colorpicker/pull/43)
+
+# 3.3.2 (2020-04-14)
+
+- Fixed the colorpicker to hide the launch and reset buttons when the input element's `type` is `hidden`. [#40](https://github.com/blackbaud/skyux-colorpicker/pull/40)
+
+# 3.3.1 (2020-04-13)
+
+- Fixed the colorpicker component to properly set ARIA attributes when the colorpicker is closed. [#38](https://github.com/blackbaud/skyux-colorpicker/pull/38)
+
+# 3.3.0 (2020-04-08)
+
+- Updated the colorpicker component to implement the affix and overlay services. [#35](https://github.com/blackbaud/skyux-colorpicker/pull/35)
+
+# 3.2.1 (2020-02-13)
+
+- Fixed the colorpicker component to use the `Renderer2` service instead of the deprecated `Renderer` service. [#50](https://github.com/blackbaud/skyux-lookup/pull/50)
+
+# 3.2.0 (2019-11-15)
+
+- Added a close message to the colorpicker component's message stream. [#29](https://github.com/blackbaud/skyux-colorpicker/pull/29) (Thanks, [@blackbaud-conorwright](https://github.com/blackbaud-conorwright)!)
+
 # 3.1.1 (2019-09-27)
 
 - Fixed the colorpicker component to properly export all output types. [#24](https://github.com/blackbaud/skyux-colorpicker/pull/24)